Refactor GetEarliestReservation to metrics route
