+++ server/db/controllers/metrics.js
@@ -1,0 +1,11 @@
+const knex = require('../dbConnections.js');
+
+const getEarliestReservationDate = async () => {
+  return await knex('Reservations').min('EventDate', {
+    as: 'EarliestReservationDate',
+  });
+};
+
+module.exports = {
+  getEarliestReservationDate,
+};
